# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FREGE AI – FregeOS Agricultural Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /crops:
    get:
      operationId: getCrops
      tags: [crops]
      summary: List all crops
      responses:
        "200":
          description: Crops list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Crop"
    post:
      operationId: createCrop
      tags: [crops]
      summary: Add a new crop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CropInput"
      responses:
        "201":
          description: Created crop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crop"
  /crops/{cropId}:
    get:
      operationId: getCrop
      tags: [crops]
      summary: Get crop details
      parameters:
        - in: path
          name: cropId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Crop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crop"
    patch:
      operationId: updateCrop
      tags: [crops]
      summary: Update crop record
      parameters:
        - in: path
          name: cropId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CropUpdate"
      responses:
        "200":
          description: Updated crop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crop"
    delete:
      operationId: deleteCrop
      tags: [crops]
      summary: Delete crop
      parameters:
        - in: path
          name: cropId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── LIVESTOCK ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Crop:
      type: object
      required: [id, farmId, name, variety, stage, plantingDate, healthStatus, healthScore, expectedYieldKg, createdAt]
      properties:
        id:
          type: integer
        farmId:
          type: integer
        name:
          type: string
        variety:
          type: ["string", "null"]
        stage:
          type: string
          enum: [planted, germination, vegetative, flowering, fruiting, harvest, harvested]
        plantingDate:
          type: string
        expectedHarvestDate:
          type: ["string", "null"]
        healthStatus:
          type: string
          enum: [excellent, good, fair, poor, critical]
        healthScore:
          type: number
        plotSizeHectares:
          type: ["number", "null"]
        expectedYieldKg:
          type: number
        actualYieldKg:
          type: ["number", "null"]
        fertilizerSchedule:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CropInput:
      type: object
      required: [farmId, name, stage, plantingDate, expectedYieldKg]
      properties:
        farmId:
          type: integer
        name:
          type: string
        variety:
          type: string
        stage:
          type: string
          enum: [planted, germination, vegetative, flowering, fruiting, harvest, harvested]
        plantingDate:
          type: string
        expectedHarvestDate:
          type: string
        plotSizeHectares:
          type: number
        expectedYieldKg:
          type: number
        fertilizerSchedule:
          type: string
        notes:
          type: string
    CropUpdate:
      type: object
      properties:
        stage:
          type: string
          enum: [planted, germination, vegetative, flowering, fruiting, harvest, harvested]
        healthStatus:
          type: string
          enum: [excellent, good, fair, poor, critical]
        healthScore:
          type: number
        actualYieldKg:
          type: number
        notes:
          type: string

    # ─── LIVESTOCK ────────────────────────────────────────────────────────────
...